+++ mayan/apps/documents/apps.py
@@ -46,39 +46,50 @@ from .handlers import (
     handler_create_default_document_type, handler_create_document_cache,
     handler_remove_empty_duplicates_lists, handler_scan_duplicates_for
 )
-from .links import (
-    link_document_clear_transformations,
-    link_document_clone_transformations, link_document_delete,
+from .links.document_links import (
+    link_document_clear_transformations, link_document_clone_transformations,
     link_document_document_type_edit, link_document_download,
-    link_document_duplicates_list, link_document_edit,
-    link_document_favorites_add, link_document_favorites_remove,
-    link_document_list, link_document_list_deleted,
-    link_document_list_favorites, link_document_list_recent_access,
-    link_document_list_recent_added,
+    link_document_duplicates_list, link_document_edit, link_document_list,
+    link_document_list_recent_access, link_document_list_recent_added,
     link_document_multiple_clear_transformations,
-    link_document_multiple_delete, link_document_multiple_document_type_edit,
-    link_document_multiple_download, link_document_multiple_favorites_add,
-    link_document_multiple_favorites_remove, link_document_multiple_restore,
-    link_document_multiple_trash, link_document_multiple_update_page_count,
-    link_document_page_disable, link_document_page_multiple_disable,
-    link_document_page_enable, link_document_page_multiple_enable,
-    link_document_page_navigation_first, link_document_page_navigation_last,
-    link_document_page_navigation_next, link_document_page_navigation_previous,
-    link_document_page_return, link_document_page_rotate_left,
-    link_document_page_rotate_right, link_document_page_view,
-    link_document_page_view_reset, link_document_page_zoom_in,
-    link_document_page_zoom_out, link_document_pages, link_document_preview,
-    link_document_print, link_document_properties, link_document_quick_download,
-    link_document_restore, link_document_trash, link_document_type_create,
-    link_document_type_delete, link_document_type_edit,
-    link_document_type_filename_create, link_document_type_filename_delete,
-    link_document_type_filename_edit, link_document_type_filename_list,
-    link_document_type_list, link_document_type_policies,
-    link_document_type_setup, link_document_update_page_count,
+    link_document_multiple_document_type_edit,
+    link_document_multiple_download, link_document_preview,
+    link_document_print, link_document_properties,
+    link_document_quick_download, link_duplicated_document_list,
+    link_duplicated_document_scan
+)
+from .links.document_version_links import (
     link_document_version_download, link_document_version_list,
     link_document_version_return_document, link_document_version_return_list,
-    link_document_version_revert, link_document_version_view,
-    link_duplicated_document_list, link_duplicated_document_scan,
+    link_document_version_revert, link_document_version_view
+)
+from .links.document_type_links import (
+    link_document_type_create, link_document_type_delete,
+    link_document_type_edit, link_document_type_filename_create,
+    link_document_type_filename_delete, link_document_type_filename_edit,
+    link_document_type_filename_list, link_document_type_list,
+    link_document_type_policies, link_document_type_setup
+)
+from .links.document_version_page_links import (
+    link_document_multiple_update_page_count, link_document_page_disable,
+    link_document_page_enable, link_document_page_multiple_enable,
+    link_document_page_multiple_disable, link_document_page_navigation_first,
+    link_document_page_navigation_last, link_document_page_navigation_next,
+    link_document_page_navigation_previous, link_document_page_return,
+    link_document_page_rotate_left, link_document_page_rotate_right,
+    link_document_page_view, link_document_page_view_reset,
+    link_document_page_zoom_in, link_document_page_zoom_out,
+    link_document_pages, link_document_update_page_count
+)
+from .links.favorite_links import (
+    link_document_favorites_add, link_document_favorites_remove,
+    link_document_list_favorites, link_document_multiple_favorites_add,
+    link_document_multiple_favorites_remove
+)
+from .links.trashed_document_links import (
+    link_document_delete, link_document_list_deleted,
+    link_document_multiple_delete, link_document_multiple_restore,
+    link_document_multiple_trash, link_document_restore, link_document_trash,
     link_trash_can_empty
 )
 from .menus import menu_documents