+++ contrib/scripts/install/docker.sh
@@ -20,6 +20,7 @@ set -e
 : ${DOCKER_POSTGRES_CONTAINER:=mayan-edms-postgres}
 : ${DOCKER_POSTGRES_VOLUME:=/docker-volumes/mayan-edms/postgres}
 : ${DOCKER_POSTGRES_PORT:=5432}
+: ${DOCKER_POSTGRES_DELAY:=10}
 : ${DOCKER_MAYAN_IMAGE:=mayanedms/mayanedms:latest}
 : ${DOCKER_MAYAN_CONTAINER:=mayan-edms}
 : ${DOCKER_MAYAN_VOLUME:=/docker-volumes/mayan-edms/media}
@@ -51,6 +52,7 @@ echo "DOCKER_POSTGRES_IMAGE: $DOCKER_POSTGRES_IMAGE"
 echo "DOCKER_POSTGRES_CONTAINER: $DOCKER_POSTGRES_CONTAINER"
 echo "DOCKER_POSTGRES_VOLUME: $DOCKER_POSTGRES_VOLUME"
 echo "DOCKER_POSTGRES_PORT: $DOCKER_POSTGRES_PORT"
+echo "DOCKER_POSTGRES_DELAY: $DOCKER_POSTGRES_DELAY"
 echo "DOCKER_MAYAN_IMAGE: $DOCKER_MAYAN_IMAGE"
 echo "DOCKER_MAYAN_CONTAINER: $DOCKER_MAYAN_CONTAINER"
 echo "DOCKER_MAYAN_VOLUME: $DOCKER_MAYAN_VOLUME"
@@ -106,8 +108,8 @@ docker run -d \
 $DOCKER_POSTGRES_IMAGE >/dev/null
 echo "Done"
 
-echo -n "* Waiting for the PostgreSQL container to be ready (10 seconds)..."
-sleep 10
+echo -n "* Waiting for the PostgreSQL container to be ready (${DOCKER_POSTGRES_DELAY} seconds)..."
+sleep $DOCKER_POSTGRES_DELAY
 echo "Done"
 
 echo -n "* Deploying Mayan EDMS container..."