+++ mayan/__init__.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 __title__ = 'Mayan EDMS'
 __version__ = '3.2.8'
 __build__ = 0x030208
-__build_string__ = 'v3.2.8_Tue Oct 1 13:31:40 2019 -0400'
+__build_string__ = 'v3.2.8-20-g2edacc9a06_Mon Oct 28 23:12:39 2019 -0400'
 __django_version__ = '1.11'
 __author__ = 'Roberto Rosario'
 __author_email__ = 'roberto.rosario@mayan-edms.com'